modules/reasoning-intervention/requirements.txt
requests==2.31.0
Flask==3.0.3
rdflib==7.0.0
SPARQLwrapper==2.0.0
google-genai==1.5.0
pydantic>=2.0